Final merge updates
